+++ lib/src/settings.rs
@@ -46,6 +46,7 @@ pub struct RepoSettings {
 pub struct GitSettings {
     pub auto_local_bookmark: bool,
     pub abandon_unreachable_commits: bool,
+    pub colocate: bool,
 }
 
 impl GitSettings {
@@ -55,6 +56,7 @@ impl GitSettings {
             abandon_unreachable_commits: config
                 .get_bool("git.abandon-unreachable-commits")
                 .unwrap_or(true),
+            colocate: config.get_bool("git.colocate").unwrap_or(false),
         }
     }
 }
@@ -64,6 +66,7 @@ impl Default for GitSettings {
         GitSettings {
             auto_local_bookmark: false,
             abandon_unreachable_commits: true,
+            colocate: false,
         }
     }
 }